Macro updates
* separate extraction of objects from RelVal
  --> now there are 2 macros

  * extraction: ExtractAndFlatten.C
    Producing what was previously "newfile1.root"/"newfile2.root".
    These contain the objects to be compared.

  * RelVal: ReleaseValidation.C
    This does the actual validation based on the extracted objects.

  This allows to extract any number of files an merge them together in
  one single file before comparsion.
  (That was previously not possible since very long cmd lines could not
  be processed by ROOT)
  In addition it is now possible to have a baseline file of objects to
  compare to without the need of extracting the baseline over and over
  again.
